blockchain-network/smart-contracts/python/requirements.txt
web3==6.20.0
python-dotenv==1.0.1
pytest==8.3.3